# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for ItsPlanned - A Collaborative Event Planning Application. This API provides endpoints for managing events, tasks, user profiles, and integrations with external services like Google Calendar.",
        "title": "ItsPlanned API",
        "contact": {
            "name": "ItsPlanned Support",
            "url": "https://github.com/vl4ddos/itsplanned",
            "email": "support@itsplanned.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/task-status-events/unread": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all unread task status events for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get unread task status events",
                "responses": {
                    "200": {
                        "description": "Unread task status events retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.TaskStatusEventsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "user": {
                    "$ref": "#/definitions/api.UserResponse"
                }
            }
        },
...
        "api.TaskStatusEventResponse": {
            "type": "object",
            "properties": {
                "changed_by_id": {
                    "type": "integer",
                    "example": 2
                },
                "changed_by_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "event_time": {
                    "type": "string",
                    "example": "2024-03-16T12:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_read": {
                    "type": "boolean",
                    "example": false
                },
                "new_status": {
                    "type": "string",
                    "example": "assigned"
                },
                "old_status": {
                    "type": "string",
                    "example": "unassigned"
                },
                "task_id": {
                    "type": "integer",
                    "example": 5
                },
                "task_name": {
                    "type": "string",
                    "example": "Buy decorations"
                }
            }
        },
        "api.TaskStatusEventsResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TaskStatusEventResponse"
                    }
                }
            }
        },
...
        "api.UserResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "bio": {
                    "type": "string",
                    "example": "Software developer and tech enthusiast"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-03-16T12:00:00Z"
                },
                "display_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-03-16T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}